% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_linear_model.R
\name{\%r\%}
\alias{\%r\%}
\title{Fitted linear model}
\usage{
y \%r\% x
}
\arguments{
\item{y}{output}

\item{x}{input}
}
\value{
fitted coefficients
}
\description{
Fitted linear model
}
\examples{

1:2 \%r\% 2:3

}
