% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\name{Person}
\alias{Person}
\title{Person object}
\usage{
Person(name, age)
}
\arguments{
\item{name}{Name of a Person}

\item{age}{Age of a Person}
}
\value{
An object of class `Person`
}
\description{
Person object
}
\examples{
Person(name = "Alice", age = 30)

}
