% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.person.R
\name{print.person}
\alias{print.person}
\title{Print method for a class person}
\usage{
\method{print}{person}(x, ...)
}
\arguments{
\item{x}{an object to which we apply our method}

\item{...}{Additional parameters (ignored)}
}
\value{
# prints information about the person
}
\description{
Print method for a class person
}
\examples{

p <- person("Alice", 30)
print(p)

}
