% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_method.R
\name{print}
\alias{print}
\title{Register a method print}
\usage{
print(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{}
}
\value{
a method
}
\description{
Register a method print
}
\examples{
print.person

}
