% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\name{person}
\alias{person}
\title{Person object}
\usage{
person(name, age)
}
\arguments{
\item{name}{Name of a person}

\item{age}{Age of a person}
}
\value{
An object of class `person`.
}
\description{
Person object
}
\examples{

person("Alice", 30)
class(p)

}
