% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Person}
\alias{print.Person}
\title{Print method for a class person}
\usage{
\method{print}{Person}(x, ...)
}
\arguments{
\item{x}{an object to which we apply our method}

\item{...}{Additional parameters (ignored)}
}
\value{
Prints information about the person
}
\description{
Print method for a class person
}
\examples{
p <- Person("Alice", 30)
print(p)

bob_object <- Person(name = "Bob", age = 27)
print(bob_object)
}
